% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actor.R
\name{Actor}
\alias{Actor}
\alias{get_actor_value}
\alias{get_actor_data}
\alias{put_actor_value}
\alias{put_actor_data}
\alias{get_actor_keys}
\title{Read and write Actor data}
\usage{
get_actor_value(id, db = default_db())

get_actor_data(ids, db = default_db())

put_actor_value(value, dimension, db = default_db())

put_actor_data(values, dimension, db = default_db())

get_actor_keys(db = default_db())
}
\arguments{
\item{id}{A UniqueID.}

\item{db}{A bedrockdb object.}

\item{ids}{A vector of UniqueIDs.}

\item{value}{An NBT object.}

\item{dimension}{An integer.}

\item{values}{A list of NBT objects.}
}
\description{
The nbt data of a actor is saved in the database, using a key with a prefix
and a 16-character storage key: 'actor:0123456789abcdef'.
}
