% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actor.R
\name{Actors}
\alias{Actors}
\alias{get_actors_data}
\alias{get_actors_value}
\title{Read and write Actor data}
\usage{
get_actors_data(x, z, dimension, db)

get_actors_value(x, z, dimension, db)
}
\arguments{
\item{x, z, dimension}{Chunk coordinates to extract data from.
\code{x} can also be a character vector of db keys.
#@param value A character vector.
#@param values A list of character vectors. If named, the names represent db
#keys.
#@param rawdata A raw vector.}

\item{db}{A bedrockdb object.}
}
\description{
Read and write Actor data
}
