% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{bedrock_random_create_seed}
\alias{bedrock_random_create_seed}
\alias{bedrock_random_create_seed1}
\title{Random Number Seeds for Minecraft}
\usage{
bedrock_random_create_seed1(x, z, a, b, salt)
}
\arguments{
\item{x, z}{chunk coordinates}

\item{a, b}{seed parameters}

\item{salt}{seed parameter}
}
\description{
\code{bedrock_random_create_seed1()} constructs a seed using the formula
\code{x*a ^ z*b ^ salt}
}
\details{
Minecraft uses several different kind of seeds during world generation and
gameplay.
}
\examples{
# identify slime chunks
g <- tidyr::expand_grid(x=1:10, z=1:10)
is_slime_chunk <- purrr::pmap_lgl(g, function(x,z) {
  seed <- bedrock_random_create_seed1(x,z,0x1f1f1f1f,1,0)
  bedrock_random_seed(seed)
  bedrock_random_get_uint(1,10) == 0
})
}
