% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomes.R
\docType{data}
\name{biome_df}
\alias{biome_df}
\title{Bedrock biome data}
\format{
A data.frame with 87 rows and 5 columns.

\describe{
\item{bedrock_id}{The numeric id of the biome.}
\item{bedrock_name}{The name of the biome.}
\item{java_name}{The name of the equivalent biome in Java edition.}
\item{universal_name}{The universal name used for the biome in Amulet.}
\item{color}{The color used when mapping biomes.}
}
}
\source{
\itemize{
\item \url{https://github.com/gentlegiantJGC/PyMCTranslate/}
\item \url{https://github.com/Cubitect/cubiomes/}
}
}
\usage{
biome_df
}
\description{
Information about biomes used in Bedrock edition. Generated from the
PyMCTranslate project. Colors are generated from the cubiomes project.
}
\keyword{datasets}
