% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{chrkeys_to_rawkeys}
\alias{chrkeys_to_rawkeys}
\alias{rawkeys_to_chrkeys}
\title{Convert between key types.}
\usage{
chrkeys_to_rawkeys(keys)

rawkeys_to_chrkeys(keys)
}
\arguments{
\item{keys}{a character vector of chrkeys or a list or rawkeys}
}
\value{
\code{chrkeys_to_rawkeys()} returns a list of raw vectors.

\if{html}{\out{<div class="sourceCode">}}\preformatted{    `rawkeys_to_chrkeys()` returns a character vector.
}\if{html}{\out{</div>}}
}
\description{
rbedrock represents database keys two different ways. \code{chrkeys} are a
human-readable format understood by most functions. \code{rawkeys} are used
internally, by the methods of \code{bedrockdb} objects and \verb{bedrock_leveldb_*}
functions.
}
\keyword{internal}
