% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{chunk_blocks}
\alias{chunk_blocks}
\alias{chunk_blocks<-}
\title{Extract or replace chunk blocks from an array}
\usage{
chunk_blocks(x, ..., drop = TRUE, origin = chunk_origin(x))

chunk_blocks(x, ..., origin = chunk_origin(x)) <- value
}
\arguments{
\item{x}{Object from which to extract element(s) or in which to replace
element(s).}

\item{...}{block coordinates specifying elements to extract or replace. Can
be numeric, logical, or missing. If numeric, the coordinates will be mapped
to indices unless there is a single, non-matrix argument.}

\item{drop}{if \code{TRUE} the result is coerced to the lowest possible dimension.}

\item{origin}{the origin of the chunk array, used for mapping coordinates to
indices}

\item{value}{An array-like R object of similar class as x}
}
\description{
Convenience wrappers around \code{[} to extract or replace blocks from an array
based on block coordinates.
}
