% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{chunk_origin}
\alias{chunk_origin}
\alias{chunk_origin<-}
\title{Get or set the coordinates of the origin of a chunk}
\usage{
chunk_origin(x)

chunk_origin(x) <- value
}
\arguments{
\item{x}{an array of block data}

\item{value}{an integer vector}
}
\description{
Get or set the coordinates of the origin of a chunk
}
