% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{default_db}
\alias{default_db}
\title{Get/set the default db connection.}
\usage{
default_db(db, check = TRUE)
}
\arguments{
\item{db}{A \code{bedrockdb} object.}

\item{check}{A \code{logical} specifying whether to check validity.}
}
\value{
For \code{default_db()}, the calculated value of the default db.
For \code{default_db(db)}, the previously manually set value of \code{default_db()}.
}
\description{
The default db is the db connection that rbedrock uses by default. It
defaults to the most recently opened db, but can also be set by the user.
}
\details{
Invoking \code{default_db()} returns the current default connection or the most
recently opened one. Invoking \code{default_db(db)} updates the current default
and returns the previous set value. \code{default_db(NULL)} can be used to unset
the default db and revert to the last opened one.
}
\examples{
dbpath <- rbedrock_example_world("example1.mcworld")
dbz <- bedrockdb(dbpath)
default_db(dbz) # set default
default_db() # returns dbz
default_db(NULL) # unset default
#cleanup
close(dbz)
unlink(dbpath, recursive = TRUE)

}
