% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{delete_data}
\alias{delete_data}
\alias{delete_value}
\title{Remove values from a bedrockdb.}
\usage{
delete_data(keys, db, writeoptions = NULL)

delete_value(key, db, writeoptions = NULL)
}
\arguments{
\item{keys}{A character vector of keys.}

\item{db}{A \code{bedrockdb} object}

\item{writeoptions}{A \code{bedrock_leveldb_writeoptions} object}

\item{key}{A scalar character.}
}
\description{
Remove values from a bedrockdb.
}
