% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_uuid}
\alias{generate_uuid}
\title{Generate random UUIDs}
\usage{
generate_uuid(n = 1L)
}
\arguments{
\item{n}{Number of ids to generate}
}
\value{
A character vector containing random (version 4) UUIDs.
}
\description{
Generate random UUIDs
}
