% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps2d.R
\name{get_2dmaps}
\alias{get_2dmaps}
\alias{read_2dmaps_data}
\alias{read_2dmaps_value}
\alias{write_2dmaps_data}
\alias{write_2dmaps_value}
\title{Read and Write 2dmaps data.}
\usage{
get_2dmaps(db, x, z, dimension)

read_2dmaps_data(object)

read_2dmaps_value(rawdata)

write_2dmaps_data(object)

write_2dmaps_value(height_map, biome_map)
}
\arguments{
\item{db}{A bedrockdb object.}

\item{x, z, dimension}{Chunk coordinates to extract 2dMaps data from.
\code{x} can also be a character vector of db keys and any keys not
representing 2dMaps data will be silently dropped.}

\item{object}{For \code{read_2dmaps_data} a list of rawdata.
For \code{write_2dmaps_data} a list of lists containing "height_map" and "biome_map" values.}

\item{rawdata}{A raw vector.}

\item{height_map}{If \code{biome_map} is missing, a \code{list} with both "height_map"
and "biome_map" values. Otherwise, a 16x16 array of height data for a chunk.}

\item{biome_map}{A 16x16 array of biome data for a chunk}
}
\description{
\code{get_2dmaps} retrieves a 2dMap data from a \code{bedrockdb}.

\code{read_2dmaps_data} parses a list of raw vector of 2dMaps data.

\code{read_2dmaps_value} parses a raw vector of 2dMaps data.

\code{write_2dmaps_data} converts list of 2dMaps data into a list of raw vector.

\code{write_2dmaps_value} converts 2dMaps data into a raw vector.
}
\examples{
heights <- matrix(63,16,16)
biomes <- matrix(1,16,16)
# Pass heights and biomes as separate parameters
write_2dmaps_value(heights, biomes)
# Pass them as a list.
obj <- list(height_map = heights, biome_map = biomes)
write_2dmaps_value(obj)

}
