% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps2d.R
\name{get_biomes}
\alias{get_biomes}
\title{Get biome information from a world.}
\usage{
get_biomes(db, x, z, dimension, return_names = TRUE)
}
\arguments{
\item{db}{A bedrockdb object.}

\item{x, z, dimension}{Chunk coordinates to extract HSA data from.
x can also be a character vector of db keys and any keys not
representing biome data will be silently dropped.}

\item{return_names}{if set to true, this will return character vectors containing
the names of each biome instead of biome ids.}
}
\value{
An list of arrays containing biome information of each key.
}
\description{
Get biome information from a world.
}
