% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_entities.R
\name{get_block_entities_data}
\alias{get_block_entities_data}
\alias{put_block_entities_data}
\title{Load and store BlockEntities NBT data}
\usage{
get_block_entities_data(db, x = get_keys(db), z, dimension)

put_block_entities_data(db, data)
}
\arguments{
\item{db}{A bedrockdb object.}

\item{x, z, dimension}{Chunk coordinates to extract data from.
\code{x} can also be a character vector of db keys and any keys not
representing BlockEntities data will be silently dropped.}

\item{data}{A named-list of key-value pairs for BlockEntities NBT data.}
}
\description{
\code{get_block_entities_data} loads BlockEntities NBT data from a \code{bedrockdb}.

\code{put_block_entities_data} stores BlockEntities NBT data into a \code{bedrockdb}.
}
