% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{get_blocks}
\alias{get_blocks}
\alias{get_subchunk_blocks}
\alias{get_subchunk}
\alias{read_subchunk}
\alias{read_subchunk_blocks}
\alias{block_palette}
\alias{subchunk_coords}
\title{Get block information stored in subchunks.}
\usage{
get_subchunk_blocks(
  db,
  x,
  z,
  dimension,
  subchunk,
  names_only = FALSE,
  extra_block = FALSE
)

get_subchunk(db, x, z, dimension, subchunk, storage = 1, simplify = TRUE)

read_subchunk(rawval, storage = 1, simplify = TRUE)

read_subchunk_blocks(rawval, names_only = FALSE, extra_block = FALSE)

block_palette(object)

subchunk_coords(ind, origins = subchunk_origins(names(ind)))
}
\arguments{
\item{db}{A bedrockdb object.}

\item{x, z, dimension, subchunk}{Subchunk Coordinates to extract block data from.
x can also be a character vector of db keys and any keys not
representing subchunk data will be silently dropped.}

\item{names_only}{Return only the names of the blocks, ignoring block states.}

\item{extra_block}{Append the extra block layer to the output (separated by ';'').
This is mostly useful if you have waterlogged blocks. If the extra block is air,
it will not be appended.}

\item{storage}{Which storage layer(s) to get. If \code{storage = NULL}, all storage layers will be
returned.}

\item{simplify}{Return a single storage layer instead of a list containing a single storage layer.}

\item{rawval}{A raw vector containing binary subchunk data}

\item{object}{A storage layer returned from \code{read_subchunk} or \code{get_subchunk}.}

\item{ind}{Numeric vector or a named list of numeric vectors containing indexes for blocks in a subchunk.}

\item{origins}{A matrix of subchunk origins.}
}
\value{
\code{get_subchunk_blocks} returns a list of character arrays.
\code{get_subchunk} and \code{read_subchunk} return a list of numeric arrays or a single array.
Each array contains block ids with a block palette stored in attribute 'palette'.
\code{palette} returns the palette.
}
\description{
Reads block information from a bedrockdb. \code{get_subchunk_blocks} reads data from multiple
subchunks and returns a named list containing a three-dimensional character array with
axes \code{x}, \code{y}, and \code{z}.
Each block is represented by a string containing block name and block states in the format
\code{blockname@state1=value1@state2=value2...}. Blocks may have 0 or mode states.
}
