% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{key_prefix}
\alias{key_prefix}
\alias{starts_with}
\alias{get_data}
\alias{get_value}
\alias{has_values}
\title{Read values stored in a bedrockdb.}
\usage{
key_prefix(prefix)

starts_with(prefix)

get_data(keys, db = default_db(), readoptions = NULL)

get_value(key, db = default_db(), readoptions = NULL)

has_values(keys, db = default_db(), readoptions = NULL)
}
\arguments{
\item{prefix}{A string specifying key prefix}

\item{keys}{A character vector of keys}

\item{db}{A \code{bedrockdb} object}

\item{readoptions}{A \code{bedrock_leveldb_readoptions} object}

\item{key}{A single key}
}
\value{
get_data()\verb{returns a named-list of raw vectors.}get_value()` returns a raw vector.

\code{has_values()} returns a logical vector.
}
\description{
Read values stored in a bedrockdb.
}
