% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{get_entities_data}
\alias{get_entities_data}
\alias{put_entities_data}
\title{Load and store Entities NBT data}
\usage{
get_entities_data(db, x = get_keys(db), z, dimension)

put_entities_data(db, data)
}
\arguments{
\item{db}{A bedrockdb object.}

\item{x, z, dimension}{Chunk coordinates to extract data from.
\code{x} can also be a character vector of db keys and any keys not
representing Entities data will be silently dropped.}

\item{data}{A named-list of key-value pairs for Entities NBT data.}
}
\description{
\code{get_entities_data} loads Entities NBT data from a \code{bedrockdb}.

\code{put_entities_data} stores Entities NBT data into a \code{bedrockdb}.
}
