% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa.R
\name{get_hsa}
\alias{get_hsa}
\title{Get HardcodedSpawnArea information from a world.}
\usage{
get_hsa(db, keys = db$keys())
}
\arguments{
\item{db}{A bedrockdb object.}

\item{keys}{A character vector of db keys to extract HSAs from.
Any keys not representing HSA data will be dropped.}
}
\value{
A table containing HSA and spawn-spot information.
}
\description{
Get HardcodedSpawnArea information from a world.
}
\examples{
db <- bedrockdb("x7fuXRc8AAA=")
hsa <- get_hsa(db)
db$close()

}
