% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pending_block_ticks.R
\name{get_pending_block_ticks_data}
\alias{get_pending_block_ticks_data}
\alias{put_pending_block_ticks_data}
\title{Load and store PendingBlockTicks NBT data}
\usage{
get_pending_block_ticks_data(db, x = get_keys(db), z, dimension)

put_pending_block_ticks_data(db, data)
}
\arguments{
\item{db}{A bedrockdb object.}

\item{x, z, dimension}{Chunk coordinates to extract data from.
\code{x} can also be a character vector of db keys and any keys not
representing PendingBlockTicks data will be silently dropped.}

\item{data}{A named-list of key-value pairs for PendingBlockTicks NBT data.}
}
\description{
\code{get_pending_block_ticks_data} loads PendingBlockTicks NBT data from a \code{bedrockdb}.

\code{put_pending_block_ticks_data} stores PendingBlockTicks NBT data into a \code{bedrockdb}.
}
