% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_block_ticks.R
\name{get_random_block_ticks_data}
\alias{get_random_block_ticks_data}
\alias{put_random_block_ticks_data}
\title{Load and store RandomBlockTicks NBT data}
\usage{
get_random_block_ticks_data(db, x = get_keys(db), z, dimension)

put_random_block_ticks_data(db, data)
}
\arguments{
\item{db}{A bedrockdb object.}

\item{x, z, dimension}{Chunk coordinates to extract data from.
\code{x} can also be a character vector of db keys and any keys not
representing RandomBlockTicks data will be silently dropped.}

\item{data}{A named-list of key-value pairs for RandomBlockTicks NBT data.}
}
\description{
\code{get_random_block_ticks_data} loads RandomBlockTicks NBT data from a \code{bedrockdb}.

\code{put_random_block_ticks_data} stores RandomBlockTicks NBT data into a \code{bedrockdb}.
}
