% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_values}
\alias{get_values}
\alias{get_value}
\title{Read values stored in a bedrockdb.}
\usage{
get_values(db, keys, readoptions = NULL)

get_value(db, key, readoptions = NULL)
}
\arguments{
\item{db}{A \code{bedrockdb} object}

\item{keys}{A character vector of keys.}

\item{readoptions}{A \code{bedrock_leveldb_readoptions} object}

\item{key}{A single key.}
}
\value{
\code{get_values} returns a named-list of raw vectors.

\code{get_value} returns a raw vector.
}
\description{
Read values stored in a bedrockdb.
}
