% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbedrock.R
\name{import_world}
\alias{import_world}
\title{Import a world from an mcworld file into the minecraftWorlds directory.}
\usage{
import_world(mcworld)
}
\arguments{
\item{mcworld}{The path to an mcworld file.}
}
\description{
Import a world from an mcworld file into the minecraftWorlds directory.
}
