% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps2d.R
\name{list_biomes}
\alias{list_biomes}
\title{List Minecraft Bedrock Edition biomes.}
\usage{
list_biomes()
}
\value{
A table containing biome names and numeric ids.
}
\description{
List Minecraft Bedrock Edition biomes.
}
