% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbedrock.R
\name{list_worlds}
\alias{list_worlds}
\title{List the worlds in the minecraftWorlds directory.}
\usage{
list_worlds(worlds_dir = worlds_dir_path())
}
\arguments{
\item{worlds_dir}{The path of the minecraftWorlds directory. It defaults to the likely path.}
}
\value{
A data.frame containing information about Minecraft worlds.
}
\description{
List the worlds in the minecraftWorlds directory.
}
