% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps2d.R
\name{Maps2D}
\alias{Maps2D}
\alias{get_2dmaps_data}
\alias{get_2dmaps_value}
\alias{read_2dmaps_value}
\alias{put_2dmaps_data}
\alias{put_2dmaps_values}
\alias{put_2dmaps_value}
\alias{write_2dmaps_value}
\title{Read and write 2DMaps data}
\usage{
get_2dmaps_data(db, x, z, dimension)

get_2dmaps_value(db, x, z, dimension)

read_2dmaps_value(rawdata)

put_2dmaps_data(db, data)

put_2dmaps_values(db, x, z, dimension, values)

put_2dmaps_value(db, x, z, dimension, value)

write_2dmaps_value(height_map, biome_map)
}
\arguments{
\item{db}{A bedrockdb object.}

\item{x, z, dimension}{Chunk coordinates to extract data from.
\code{x} can also be a character vector of db keys.}

\item{rawdata}{A raw vector.}

\item{data}{A named-vector of key-value pairs for 2DMaps data.}

\item{values}{A list of lists.}

\item{value}{A list.}

\item{height_map}{If \code{biome_map} is missing, a \code{list} with both "height_map"
and "biome_map" values. Otherwise, a 16x16 array of height
data for a chunk.}

\item{biome_map}{A 16x16 array of biome data for a chunk}
}
\value{
\code{get_2dmaps_data()} returns a list of the of the values returned by
\code{get_2dmaps_value()}.

\code{get_2dmaps_value()} and \code{read_2dmaps_value()}
return a list with components "height_map" and "biome_map".
}
\description{
2DMaps data (tag 45) stores information about surface heights and biomes in a
chunk. 2DMaps data is 768 bytes long and consists of a 256 int16s (heights)
followed by 256 uint8s (biomes).

\code{get_2dmaps_data()} loads 2DMaps data from a \code{bedrockdb}.
It will silently drop and keys not representing 2DMaps data.

\code{get_2dmaps_value()} loads 2DMaps data from a \code{bedrockdb}.
It only supports loading a single value.

\code{read_2dmaps_value} decodes binary 2DMaps data.

\code{put_2dmaps_data()}, \code{put_2dmaps_values()}, and
\code{put_2dmaps_value()} store 2DMaps data into a \code{bedrockdb}.

\code{write_2dmaps_value} encodes 2dMaps data into a raw vector.
}
\examples{
heights <- matrix(63,16,16)
biomes <- matrix(1,16,16)
# Pass heights and biomes as separate parameters
write_2dmaps_value(heights, biomes)
# Pass them as a list.
obj <- list(height_map = heights, biome_map = biomes)
write_2dmaps_value(obj)

}
