% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbedrock.R
\name{minecraft_worlds}
\alias{minecraft_worlds}
\alias{worlds_dir_path}
\alias{list_worlds}
\alias{export_world}
\alias{import_world}
\title{Utilities for working with Minecraft world folders.}
\usage{
worlds_dir_path(default = TRUE)

list_worlds(worlds_dir = worlds_dir_path())

export_world(world_folder, mcworld_path, worlds_dir = worlds_dir_path())

import_world(mcworld_path, worlds_dir = worlds_dir_path())
}
\arguments{
\item{default}{If \code{TRUE}, return most likely world path on the system.}

\item{worlds_dir}{The path of a \code{minecraftWorlds} directory.}

\item{world_folder}{The path to a world folder. If the path does not exist, it is 
assumed to be the base name of a world folder in \code{worlds_dir}.}

\item{mcworld_path}{The path to an mcworld file. If exporting, it will be created
and overwritten if it exists. If importing, it will be extracted.}
}
\value{
\code{world_dir_path} returns the path to the \code{minecraftWorlds} directory.
        You can use \code{options(rbedrock.worlds_dir_path = "custom/path")} to customize the path
        as needed.
        \code{list_worlds} returns a data.frame containing information about Minecraft saved games.
}
\description{
Utilities for working with Minecraft world folders.
}
