% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbt-build.R
\name{nbt_build}
\alias{nbt_build}
\alias{nbt_build_compound}
\title{Convert a basic R object into an NBT object}
\usage{
nbt_build(x)

nbt_build_compound(...)
}
\arguments{
\item{x}{an R object}

\item{...}{R objects, possibly named.}
}
\value{
An NBT object with type inferred from x. \code{nbt_build_compound()}
always returns an nbt_compound constructed from \code{...}.
}
\description{
The Named Binary Tag (NBT) format is used by Minecraft for various data
types. An NBT value holds a 'payload' of data and a 'tag' indicating the type
of data held. The "build" functions are useful for automatically converting
basic R objects into the most commonly used NBT objects.
}
\details{
The following conversion rules are used:
\itemize{
\item nbt_byte: logical or raw scalars
\item nbt_short: complex scalars (via \code{Im(x)})
\item nbt_int: integer scalars
\item nbt_long: scalars of class integer64
\item nbt_float: double scalars
\item nbt_string: character scalars
\item nbt_byte_list: logical vectors
\item nbt_raw_string: raw vectors
\item nbt_short_list: complex vectors (via \code{Im(x)})
\item nbt_int_list: integer vectors
\item nbt_long_list: vectors of class integer64
\item nbt_float_list: double vectors
\item nbt_string_list: character vectors
\item nbt_empty_list: NULL
\item nbt_compound: named lists
\item nbt_compound_list: unnamed lists
}

In addition, \code{I()} can be used to treat scalars as vectors of length 1, and
to convert unnamed lists to nbt_compounds.
}
\examples{
LL <- bit64::as.integer64(1L)
nbt_build_compound(
  byte = TRUE,
  short = 1i,
  int = 1L,
  long = 1 * LL,
  float = 1.0,
  string = "1"
)
}
\seealso{
\link{nbt} \link[base:AsIs]{base::AsIs} \link[base:complex]{base::complex}
}
