% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbt.R
\name{nbtnode}
\alias{nbtnode}
\alias{new_nbtnode}
\title{NBTnode constructor}
\usage{
nbtnode(payload, tag, list_tag = NULL)

new_nbtnode(payload, tag, list_tag = NULL, ...)
}
\arguments{
\item{payload}{A nbtnode payload}

\item{tag}{A numeric tag for the node.}

\item{list_tag}{A numeric tag for values stored in the payload of an NBT "list" node.}

\item{...}{Additional parameters passed to \code{structure} when creating the object.}
}
\description{
\code{nbtnode} is a convenient wrapper around \code{new_nbtnode} that supports
type conversion.

\code{new_nbtnode} creates a new object with class \code{nbtnode}.
}
