% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbt.R
\name{payload}
\alias{payload}
\alias{payload<-}
\title{Read and write an \code{nbtnode}'s payload}
\usage{
payload(object)

payload(object) <- value
}
\arguments{
\item{object}{An nbtnode}

\item{value}{A new payload
'}
}
\description{
Read and write an \code{nbtnode}'s payload
}
