% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps2d.R
\name{put_biomes}
\alias{put_biomes}
\title{Put biome information into the world}
\usage{
put_biomes(db, values, x = names(values), z, dimension, missing_height = 0L)
}
\arguments{
\item{db}{A bedrockdb object.}

\item{values}{A list of character or integer vectors. Each element of
the list must contain 256 values or an error will be raised.}

\item{x, z, dimension}{Chunk coordinates to write biome data to.
`x` can also be a character vector of db keys and any keys not
representing biome data (tag 47) will be silently dropped.
`x` defaults to `names(values)`}

\item{missing_height}{if there is no existing height data, use this value for the chunk.}
}
\description{
put_biomes updates the biome information of chunks.
It will preserve any existing height data or use
the missing_height value if no such data exists.
}
