% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{put_data}
\alias{put_data}
\alias{put_value}
\title{Write values to a bedrockdb.}
\usage{
put_data(values, keys, db = default_db(), writeoptions = NULL)

put_value(value, key, db = default_db(), writeoptions = NULL)
}
\arguments{
\item{values}{A list of raw values.}

\item{keys}{A character vector of keys.}

\item{db}{A \code{bedrockdb} object}

\item{writeoptions}{A \code{bedrock_leveldb_writeoptions} object}

\item{value}{A raw vector that contains the information to be written.}

\item{key}{A key that will be used to store data.}
}
\value{
An invisible copy of \code{db}.
}
\description{
Write values to a bedrockdb.
}
