% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa.R
\name{put_hsa}
\alias{put_hsa}
\title{Add a HardcodedSpawnArea to a world.}
\usage{
put_hsa(db, x1, y1, z1, x2, y2, z2, tag, dimension = ifelse(tag == 1, 1, 0))
}
\arguments{
\item{db}{A bedrockdb object.}

\item{x1, y1, z1, x2, y2, z2}{HSA bounding box coordinates.}

\item{tag}{The type of HSA. 1 = NetherFortress, 2 = SwampHut, 3 = OceanMonument, 5 = PillagerOutpost.
4 and 6 are no longer used by the game.}

\item{dimension}{The dimension that the HSA should be in. 0 = Overworld, 1 = Nether.}
}
\value{
A table containing information about the added HSAs.
}
\description{
If the bounding box of the HSA overlaps multiple chunks, one HSA will be added per chunk.
}
\examples{
db <- bedrockdb("x7fuXRc8AAA=")
put_hsa(db, 0, 60, 0, 15, 70, 15, 2, 0)
db$close()

}
