% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rac_cast}
\alias{rac_cast}
\title{Cast a value to a specific type}
\usage{
rac_cast(x, to, ..., x_arg = "x")
}
\arguments{
\item{x}{Value to cast}

\item{to}{Type to cast to}

\item{...}{Currently unused.}

\item{x_arg}{Argument name for \code{x}, used in error messages.}
}
\description{
This is a helper function used (mostly) for modifying NBT values.
}
\keyword{internal}
