% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rac_type_full}
\alias{rac_type_full}
\alias{rac_type_abbr}
\title{Object type as a string}
\usage{
rac_type_full(x)

rac_type_abbr(x)
}
\arguments{
\item{x}{An object.}
}
\description{
\code{rac_type_full()} displays the fill type of the object.
\code{rac_type_abbr()} provides an abbreviated description.
}
\keyword{internal}
