% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_block_ticks.R
\name{RandomBlockTicks}
\alias{RandomBlockTicks}
\alias{get_random_block_ticks_data}
\alias{put_random_block_ticks_data}
\title{Load and store RandomBlockTicks NBT data}
\usage{
get_random_block_ticks_data(db, x = get_keys(db), z, dimension)

put_random_block_ticks_data(db, data)
}
\arguments{
\item{db}{A bedrockdb object.}

\item{x, z, dimension}{Chunk coordinates to extract data from.
\code{x} can also be a character vector of db keys.}

\item{data}{A named-list of key-value pairs for RandomBlockTicks data.}
}
\description{
RandomBlockTicks data (tag 59) holds a list of NBT values for
random ticks.

\code{get_random_block_ticks_data()} loads RandomBlockTicks data from a \code{bedrockdb}.
It will silently drop and keys not representing RandomBlockTicks data.

\code{put_random_block_ticks_data()} stores RandomBlockTicks data into a \code{bedrockdb}.
}
