% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbedrock-package.R
\docType{package}
\name{rbedrock-package}
\alias{rbedrock}
\alias{rbedrock-package}
\title{rbedrock: Analysis and Manipulation of Data from Minecraft Bedrock Edition}
\description{
This package is an extension package for R that supports the analysis
  and management of Minecraft (Bedrock Edition) worlds. This includes Windows 10, Pocket Edition, XBox, and PS4 versions of the game. It does not include Minecraft: Java worlds.
}
\details{
rbedrock is currently under development.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/reedacartwright/rbedrock/}
  \item Report bugs at \url{http://github.com/reedacartwright/rbedrock/issues}
}

}
\author{
\strong{Maintainer}: Reed Cartwright \email{racartwright@gmail.com} (\href{https://orcid.org/0000-0002-0837-9380}{ORCID})

Authors:
\itemize{
  \item Rich FitzJohn \email{rich.fitzjohn@gmail.com}
}

}
\keyword{internal}
