% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbedrock.R
\name{rbedrock}
\alias{rbedrock}
\title{rbedrock: A package for the analysis and manipulation of Minecraft:
Bedrock Edition worlds.}
\description{
rbedrock is an extension package for R that supports the analysis
and management of Minecraft (Bedrock Edition) worlds. This includes
Windows 10, Pocket Edition, XBox, and PS4 versions of the game. It
does not include Minecraft: Java worlds. @docType package
}
