% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spawning_area}
\alias{spawning_area}
\title{Calculate a player-based spawning area}
\usage{
spawning_area(sim_distance, x = 0, z = 0)
}
\arguments{
\item{sim_distance}{A sim distance setting}

\item{x, z}{Chunk coordinates where a player is standing (can be fractional)}
}
\value{
A \code{data.frame} containing the chunk coordinates in the spawning area.
}
\description{
Calculate a player-based spawning area
}
