% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbt-rnbt.R
\name{to_rnbt_value}
\alias{to_rnbt_value}
\title{Convert an nbt value to an rnbt value}
\usage{
to_rnbt_value(x, ...)
}
\arguments{
\item{x}{Value to cast}

\item{...}{Currently unused.}
}
\description{
This is a helper function used to convert from nbt values to rnbt data.
}
\keyword{internal}
