% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks-str.R
\docType{data}
\name{vanilla_block_states_df}
\alias{vanilla_block_states_df}
\alias{vanilla_block_list}
\alias{vanilla_block_property_type_list}
\title{Bedrock block data}
\format{
\subsection{vanilla_block_states_df}{

A data.frame in long-format with 1131 rows and
5 columns. Block data version is
18168865.

\describe{
\item{name}{Block name.}
\item{property}{Property name.}
\item{type}{Property type.}
\item{default}{Default value.}
\item{allowed}{Allowed values.}
}
}

\subsection{vanilla_block_list}{

List of blocks names. Includes blocks without properties, which don't show
up in vanilla_block_states_df.
}

\subsection{vanilla_block_property_type_list}{

List of properties (names) and their types (values).
}
}
\source{
\itemize{
\item \url{https://github.com/gentlegiantJGC/PyMCTranslate/}
}
}
\usage{
vanilla_block_states_df

vanilla_block_list

vanilla_block_property_type_list
}
\description{
Information about blocks used in Bedrock edition. Generated from the
PyMCTranslate project.
}
\keyword{datasets}
