% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbedrock.R
\name{worlds_path}
\alias{worlds_path}
\title{The local minecraftWorlds directory.}
\usage{
worlds_path()
}
\value{
The likely path to the local minecraftWorlds directory.
}
\description{
The local minecraftWorlds directory.
}
