% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{write_data}
\alias{write_data}
\title{Write values to a bedrockdb.}
\usage{
write_data(values, keys, db, writeoptions = NULL)
}
\arguments{
\item{values}{A list of raw values. Use \code{zap()} or \code{NULL} to indicate
that a value should be delete.
If \code{keys} is missing, the names of \code{values} will be taken as the keys.}

\item{keys}{A character vector of keys.}

\item{db}{A \code{bedrockdb} object}

\item{writeoptions}{A \code{bedrock_leveldb_writeoptions} object}
}
\value{
An invisible copy of \code{db}.
}
\description{
\code{write_data()} commits a batch of write and delete
}
