% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\name{bedrockdb}
\alias{bedrockdb}
\title{Open a Minecraft: Bedrock Edition world for reading and writing.}
\usage{
bedrockdb(
  path,
  create_if_missing = FALSE,
  error_if_exists = NULL,
  paranoid_checks = NULL,
  write_buffer_size = 4194304,
  max_open_files = NULL,
  block_size = NULL,
  use_compression = 4,
  cache_capacity = 41943040,
  bloom_filter_bits_per_key = 10
)
}
\arguments{
\item{path}{The path to a world folder. If the path does not exist, it is
assumed to be the base name of a world folder in the local minecraftWorlds
directory.}
}
\value{
On success, an R6 class of type "bedrockdb" that contains and open handle to the
leveldb database that contains the world information.
}
\description{
Open a Minecraft: Bedrock Edition world for reading and writing.
}
\examples{
\dontrun{db <- bedrockdb("x7fuXRc8AAA=")
o <- db$get("Overworld")
read_nbt(o)
db$close()}

}
