% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_entity.R
\name{BlockEntity}
\alias{BlockEntity}
\alias{get_block_entity_data}
\alias{get_block_entity_value}
\alias{put_block_entity_data}
\alias{put_block_entity_value}
\title{Load and store BlockEntity NBT data}
\usage{
get_block_entity_data(x, z, dimension, db)

get_block_entity_value(x, z, dimension, db)

put_block_entity_data(values, x, z, dimension, db)

put_block_entity_value(value, x, z, dimension, db)
}
\arguments{
\item{x, z, dimension}{Chunk coordinates to extract data from.
\code{x} can also be a character vector of db keys.}

\item{db}{A bedrockdb object.}

\item{values}{A list of nbt objects. If \code{x} is missing, the names of \code{values}
will be taken as the keys.}

\item{value}{An nbt object.}
}
\value{
\code{get_block_entity_data()} returns a named-list of nbt data.
\code{get_block_entity_value()} returns a single nbt value.
}
\description{
BlockEntity data (tag 49) holds a list of NBT values for
entity data associated with specific blocks.
}
\details{
\code{get_block_entity_data()} and \code{get_block_entity_value()} load BlockEntity
data from \code{db}. \code{get_block_entity_value()} supports loading
only a single value.

\code{put_block_entity_value()}, and \code{put_block_entity_data()} store BlockEntity
data into \code{db}.
}
