% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_version.R
\name{ChunkVersion}
\alias{ChunkVersion}
\alias{get_chunk_version_data}
\alias{get_chunk_version_value}
\alias{put_chunk_version_data}
\alias{put_chunk_version_value}
\alias{read_chunk_version_value}
\alias{write_chunk_version_value}
\title{Read and write chunk version data}
\usage{
get_chunk_version_data(x, z, dimension, db)

get_chunk_version_value(x, z, dimension, db)

put_chunk_version_data(values, x, z, dimension, db)

put_chunk_version_value(db, x, z, dimension, value)

read_chunk_version_value(rawdata)

write_chunk_version_value(num)
}
\arguments{
\item{x, z, dimension}{Chunk coordinates to extract version data from.
\code{x} can also be a character vector of db keys.}

\item{db}{A bedrockdb object.}

\item{values}{An integer vector. If \code{x} is missing, the names of \code{values} will be taken as the keys.}

\item{value}{A scalar integer vector.}

\item{rawdata}{A scalar raw.}

\item{num}{A scalar integer.}
}
\description{
Version data (tag 44) and LegacyVersion data (tag 118)
store the version number of a chunk. In Minecraft version 1.16.100,
chunk version data was moved from tag 118 to tag 44.
}
\details{
\code{get_chunk_version_data()} and \code{get_chunk_version_value()} load Version
data from \code{db}. \code{get_chunk_version_value()} supports loading
only a single value.

\code{put_chunk_version_data()} and
\code{put_chunk_version_value()} store Version data into a \code{bedrockdb}.

\code{read_chunk_version_value()} decodes Version data and
\code{write_chunk_version_value()} encodes Version data.
}
