% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{delete_values}
\alias{delete_values}
\title{Remove values from a bedrockdb.}
\usage{
delete_values(
  db,
  keys,
  report = FALSE,
  readoptions = NULL,
  writeoptions = NULL
)
}
\arguments{
\item{db}{A \code{bedrockdb} object}

\item{keys}{A character vector of keys.}

\item{report}{A logical indicating whether to generate a report on deleted
keys}

\item{readoptions}{A \code{bedrock_leveldb_readoptions} object}

\item{writeoptions}{A \code{bedrock_leveldb_writeoptions} object}
}
\value{
If \code{report == TRUE}, a logical vector indicating which keys were
deleted.
}
\description{
Remove values from a bedrockdb.
}
