% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity.R
\name{Entity}
\alias{Entity}
\alias{get_entity_data}
\alias{get_entity_value}
\alias{put_entity_data}
\alias{put_entity_value}
\title{Load and store Entity NBT data}
\usage{
get_entity_data(x, z, dimension, db)

get_entity_value(x, z, dimension, db)

put_entity_data(values, x, z, dimension, db)

put_entity_value(value, x, z, dimension, db)
}
\arguments{
\item{x, z, dimension}{Chunk coordinates to extract data from.
\code{x} can also be a character vector of db keys.}

\item{db}{A bedrockdb object.}

\item{values}{A list of nbt objects. If \code{x} is missing, the names of \code{values}
will be taken as the keys.}

\item{value}{An nbt object.}
}
\value{
\code{get_entity_data()} returns a named-list of nbt data.
\code{get_entity_value()} returns a single nbt value.
}
\description{
Entity data (tag 50) holds a list of NBT values for mobs and other entities
in the game. After 1.18.30, entity data was migrated to a new actor digest
format and no longer saved with chunk data.
}
\details{
\code{get_entity_data()} and \code{get_entity_value()} load Entity
data from \code{db}. \code{get_entity_value()} supports loading
only a single value.

\code{put_entity_value()}, and \code{put_entity_data()} store Entity
data into \code{db}.
}
