% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbedrock.R
\name{export_world}
\alias{export_world}
\alias{import_world}
\title{Export and Import Minecraft Worlds.}
\usage{
export_world(world_folder, mcworld_path, worlds_dir = worlds_dir_path())

import_world(mcworld_path, worlds_dir = worlds_dir_path())
}
\arguments{
\item{world_folder}{The path to a world folder. If the path does not exist, it is 
assumed to be the base name of a world folder in \code{worlds_dir}.
directory.}

\item{mcworld_path}{The path to an mcworld file. If exporting, it will be created
and overwritten if it exists. If importing, it will be extracted.}

\item{worlds_dir}{The path of the \code{minecraftWorlds} directory.}
}
\description{
Export and Import Minecraft Worlds.
}
