% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps2d.R
\name{get_2dmaps}
\alias{get_2dmaps}
\alias{read_2dmaps_data}
\alias{write_2dmaps_data}
\title{Read and Write 2dmaps data.}
\usage{
get_2dmaps(db, x, z, dimension)

read_2dmaps_data(rawval)

write_2dmaps_data(object, biome_map)
}
\arguments{
\item{db}{A bedrockdb object.}

\item{x, z, dimension}{Chunk coordinates to extract 2dMaps data from.
x can also be a character vector of db keys and any keys not
representing 2dMaps data will be silently dropped.}

\item{rawval}{A \code{raw} vector containing binary 2dMaps data.}

\item{object}{A list containing height and biome biome. If \code{biome_map} is specified,
this can be a 16x16 array of height data for a chunk.}

\item{biome_map}{A 16x16 array of biome data for a chunk}
}
\description{
\code{get_2dmaps} retrieves a 2dMap data from a \code{bedrockdb}.

\code{read_2dmaps_data} parses a raw vector of 2dMaps data.

\code{write_2dmaps_data} converts 2dMaps data into a raw vector.
}
