% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkversion.R
\name{get_chunk_versions}
\alias{get_chunk_versions}
\alias{read_chunk_version_data}
\alias{write_chunk_version_data}
\title{Read and write chunk version data}
\usage{
get_chunk_versions(db, x, z, dimension, include_legacy = TRUE)

read_chunk_version_data(object)

write_chunk_version_data(object)
}
\arguments{
\item{db}{A bedrockdb object.}

\item{x, z, dimension}{Chunk coordinates to extract version data from.
x can also be a character vector of db keys and any keys not
representing version data will be silently dropped.}

\item{include_legacy}{If true, \code{ChunkVersionLegacy} tags will be included.}

\item{object}{For \code{read_chunk_version_data}, this is a list of raw values.
For \code{write_chunk_version_data}, this is a list of integers.}
}
\description{
\code{get_chunk_versions} retrieves chunk versions from a \code{bedrockdb}.

\code{read_chunk_version_data} parses chunk version data.

\code{write_chunk_version_data} converts chunk version data into binary form.
}
