% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data}
\alias{get_data}
\alias{key_prefix}
\alias{get_value}
\alias{has_values}
\title{Read values stored in a bedrockdb.}
\usage{
get_data(keys, db, readoptions = NULL)

key_prefix(prefix)

get_value(key, db, readoptions = NULL)

has_values(keys, db, readoptions = NULL)
}
\arguments{
\item{keys}{A character vector of keys.}

\item{db}{A \code{bedrockdb} object}

\item{readoptions}{A \code{bedrock_leveldb_readoptions} object}

\item{prefix}{A string specifying key prefix.}

\item{key}{A single key.}
}
\value{
\code{get_data()} returns a named-list of raw vectors. \code{get_value()} returns a raw vector.

\code{has_values()} returns a logical vector.
}
\description{
Read values stored in a bedrockdb.
}
