% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa.R
\name{get_hsa}
\alias{get_hsa}
\title{Get HardcodedSpawnArea information from a world.}
\usage{
get_hsa(db, x = get_keys(db), z, dimension)
}
\arguments{
\item{db}{A bedrockdb object.}

\item{x, z, dimension}{Chunk coordinates to extract HSA data from.
x can also be a character vector of db keys and any keys not
representing HSA data will be silently dropped.}
}
\value{
A table containing HSA and spawn-spot information.
}
\description{
Get HardcodedSpawnArea information from a world.
}
\examples{
\dontrun{db <- bedrockdb("x7fuXRc8AAA=")
hsa <- get_hsa(db)
close(db)}

}
