% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_keys}
\alias{get_keys}
\title{Get a list of keys stored in a bedrockdb.}
\usage{
get_keys(prefix = NULL, db = the_db(), readoptions = NULL)
}
\arguments{
\item{prefix}{A string specifying chunk prefix or string prefix.}

\item{db}{A \code{bedrockdb} object}

\item{readoptions}{A \code{bedrock_leveldb_readoptions} object}
}
\value{
A character vector.
}
\description{
Returns a vector containing keys found in the bedrockdb
with a specific \code{prefix}. If \code{prefix} is \code{NULL}, it will
return all the keys. If \code{db} is missing, \code{get_keys()},
will check if \code{db} was passed as the first argument.
}
\examples{
dbpath <- rbedrock_example_world("example1.mcworld")
db <- bedrockdb(dbpath)
# get all keys in the world
keys <- get_keys(db = db)
# this also works
keys <- get_keys(db)
# get all the keys in the world with a prefix
keys <- get_keys("plain:VILLAGE", db)
close(db)
}
