% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_keys}
\alias{get_keys}
\title{Get a list of keys stored in a bedrockdb.}
\usage{
get_keys(db, starts_with = NULL, readoptions = NULL)
}
\arguments{
\item{db}{A \code{bedrockdb} object}

\item{starts_with}{A string specifying chunk prefix or string prefix.}

\item{readoptions}{A \code{bedrock_leveldb_readoptions} object}
}
\value{
A vector containing all the keys found in the bedrockdb.

If \code{starts_with} is specified, this vector will be filtered for
based on the specified prefix.
}
\description{
Get a list of keys stored in a bedrockdb.
}
