% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_values}
\alias{get_values}
\alias{get_data}
\alias{get_value}
\alias{has_values}
\title{Read values stored in a bedrockdb.}
\usage{
get_values(db, keys, starts_with, readoptions = NULL)

get_data(db, keys, starts_with, readoptions = NULL)

get_value(db, key, readoptions = NULL)

has_values(db, keys, readoptions = NULL)
}
\arguments{
\item{db}{A \code{bedrockdb} object}

\item{keys}{A character vector of keys.}

\item{starts_with}{A string specifying chunk prefix or string prefix.}

\item{readoptions}{A \code{bedrock_leveldb_readoptions} object}

\item{key}{A single key.}
}
\value{
\code{get_values()} returns a named-list of raw vectors.

\code{get_value()} returns a raw vector.

\code{has_values()} returns a logical vector.
}
\description{
\code{get_values()} and \code{get_data()} are synonyms.
}
