% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa.R
\name{HardcodedSpawnArea}
\alias{HardcodedSpawnArea}
\alias{get_hsa_data}
\alias{get_hsa_value}
\alias{read_hsa_value}
\alias{put_hsa_data}
\alias{put_hsa_value}
\alias{write_hsa_value}
\title{Load and store HardcodedSpawnArea (HSA) data}
\usage{
get_hsa_data(x, z, dimension, db, simplify = TRUE)

get_hsa_value(x, z, dimension, db)

read_hsa_value(rawdata, pretty = TRUE)

put_hsa_data(values, x, z, dimension, db, merge = TRUE)

put_hsa_value(value, x, z, dimension, db)

write_hsa_value(value)
}
\arguments{
\item{x, z, dimension}{Chunk coordinates to extract data from.
\code{x} can also be a character vector of db keys.}

\item{db}{A bedrockdb object.}

\item{simplify}{If \code{TRUE}, simplify a list of tables into one table.}

\item{rawdata}{A scalar raw.}

\item{pretty}{If \code{TRUE}, return as a tibble with extra information.}

\item{values}{A table or a list of tables containing HSA coordinates and tags.
If \code{values} is a table, the HSA bounding boxes will be split across chunks as needed.}

\item{merge}{Merge the new HSAs with existing HSAs.}

\item{value}{A table containing HSA coordinates}
}
\value{
\code{get_hsa_data()} returns a table in the same format
as \code{get_hsa_value()}.
\code{get_hsa_value()} and \code{read_hsa_value()}
return a table with columns indicating the
coordinates of the HSA bounding box and the
location of the HSS at the center of the bounding
box. \code{get_hsa_value()} also records the dimension
of the bounding box.
}
\description{
HardcodedSpawnArea (HSA) data (tag 57) stores information
about structure spawning locations in a chunk.
An HSA is defined by a bounding box that specifies
the location of an HSA in a chunk and a tag that
specifies the type: 1 = NetherFortress, 2 = SwampHut,
3 = OceanMonument, and 5 = PillagerOutpost.
}
\details{
\code{get_hsa_data()} and \code{get_hsa_value()} load HardcodedSpawnArea data from \code{db}.
\code{get_hsa_value()} only supports loading a single value.

\code{put_hsa_data()} and \code{put_hsa_value()} puts HSA data into \code{db}.

\code{read_hsa_value()} decodes HSA data. \code{write_hsa_value()} encodes HSA data.
}
\examples{
dbpath <- rbedrock_example_world("example1.mcworld")
db <- bedrockdb(dbpath)
# view all HSA in a world
keys <- grep(":57$", get_keys(db), value=TRUE)
hsa <- get_hsa_data(keys, db = db)
hsa
# add an HSA to a world
dat <- data.frame(x1 = 0, x2 = 15, z1 = 0, z2 = 15,
                  y1 = 40, y2 = 60, tag = "SwampHut")
put_hsa_data(dat, db = db, merge = TRUE)
close(db)
}
