% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps2d.R
\name{list_biomes}
\alias{list_biomes}
\alias{biome_id}
\title{List Minecraft Bedrock Edition biomes.}
\usage{
list_biomes()

biome_id(x)
}
\arguments{
\item{x}{A character vector containing biome name.}
}
\value{
\code{list_biomes} returns a table containing biome names and numeric ids.
        \code{biome_id} returns the numeric ids corresponding to input names.
}
\description{
List Minecraft Bedrock Edition biomes.
}
