% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomes.R
\name{list_biomes}
\alias{list_biomes}
\alias{biome_id}
\title{List Minecraft Bedrock Edition biomes.}
\usage{
list_biomes()

biome_id(x)
}
\arguments{
\item{x}{A character vector containing biome name.}
}
\description{
List Minecraft Bedrock Edition biomes.
}
