% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps3d.R
\name{Maps3D}
\alias{Maps3D}
\alias{get_3dmaps_data}
\alias{get_3dmaps_values}
\alias{get_3dmaps_value}
\alias{read_3dmaps_value}
\title{Read and write 3DMaps data}
\usage{
get_3dmaps_data(db, x, z, dimension)

get_3dmaps_values(db, x, z, dimension)

get_3dmaps_value(db, x, z, dimension)

read_3dmaps_value(rawdata)
}
\arguments{
\item{db}{A bedrockdb object.}

\item{x, z, dimension}{Chunk coordinates to extract data from.
\code{x} can also be a character vector of db keys.}

\item{rawdata}{A raw vector.}
}
\value{
\code{get_3dmaps_data()} returns a list of the of the values returned by
\code{get_3dmaps_value()}.

\code{get_3dmaps_value()} returns a list with components "height_map"
and "biome_map".
}
\description{
3DMaps data (tag 43) stores information about surface heights and biomes in a
chunk.

\code{get_3dmaps_data()} loads 3DMaps data from a \code{bedrockdb}.
It will silently drop and keys not representing 3DMaps data.

\code{get_3dmaps_value()} loads 3DMaps data from a \code{bedrockdb}.
It only supports loading a single value.

\code{read_3dmaps_value()} decodes binary 3DMaps data.
}
