% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbt.R
\name{nbt}
\alias{nbt}
\alias{nbt_byte}
\alias{nbt_short}
\alias{nbt_int}
\alias{nbt_float}
\alias{nbt_double}
\alias{nbt_byte_array}
\alias{nbt_int_array}
\title{Create an NBT value}
\usage{
nbt(...)

nbt_byte(x)

nbt_short(x)

nbt_int(x)

nbt_float(x)

nbt_double(x)

nbt_byte_array(x)

nbt_int_array(x)
}
\arguments{
\item{x}{An nbt payload.}
}
\description{
The Named Binary Tag (NBT) format is used by Minecraft for various data
types. An NBT value holds a 'payload' of data and a 'tag' indicating the type
of data held.
}
\details{
\itemize{
\item \verb{nbt_*()} family of functions create nbt data types.
\item \code{unnbt()} recursively strips NBT metadata from an NBT value.
}
}
