% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbt.R
\name{new_nbt}
\alias{new_nbt}
\title{Create an NBT value}
\usage{
new_nbt(x = list(), tag = 0L)
}
\arguments{
\item{x}{An nbt payload.}

\item{tag}{An integer specifying the tag of the data.}
}
\description{
Create an NBT value
}
\keyword{internal}
