% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R, R/keys.R
\name{subchunk_coords}
\alias{subchunk_coords}
\alias{parse_chunk_keys}
\alias{chunk_pos}
\alias{chunk_origins}
\alias{subchunk_origins}
\title{Extract information from chunk keys.}
\usage{
subchunk_coords(offsets, origins = subchunk_origins(names(offsets)))

parse_chunk_keys(keys)

chunk_pos(keys)

chunk_origins(keys)

subchunk_origins(keys)
}
\arguments{
\item{keys}{A character vector of database keys.}
}
\value{
\code{parse_chunk_keys} returns a tibble containing information extracted
from chunk keys. Keys that do not contain chunk data are dropped.
\code{chunk_pos} return a matrix containing chunk coordinates.
\code{chunk_origins} returns the block position of the NW corner of chunks.
\code{subchunk_origins} returns the block position of the lower NW corner of subchunk.
}
\description{
Extract information from chunk keys.
}
\examples{
parse_chunk_keys("@0:0:0:47-1")
}
