% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbt.R
\name{payload}
\alias{payload}
\alias{payload<-}
\title{Read and write an \code{nbt}'s payload}
\usage{
payload(x)

payload(x) <- value
}
\arguments{
\item{x}{An nbt object}

\item{value}{A new payload}
}
\description{
Read and write an \code{nbt}'s payload
}
