% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pending_ticks.R
\name{PendingTicks}
\alias{PendingTicks}
\alias{get_pending_ticks_data}
\alias{get_pending_ticks_value}
\alias{put_pending_ticks_data}
\alias{put_pending_ticks_value}
\title{Load and store PendingTicks NBT data}
\usage{
get_pending_ticks_data(x, z, dimension, db)

get_pending_ticks_value(x, z, dimension, db)

put_pending_ticks_data(values, x, z, dimension, db)

put_pending_ticks_value(value, x, z, dimension, db)
}
\arguments{
\item{x, z, dimension}{Chunk coordinates to extract data from.
\code{x} can also be a character vector of db keys.}

\item{db}{A bedrockdb object.}

\item{values}{A list of nbt objects. If \code{x} is missing, the names of \code{values}
will be taken as the keys.}

\item{value}{An nbt object.}
}
\value{
\code{get_pending_ticks_data()} returns a named-list of nbt data.
\code{get_pending_ticks_value()} returns a single nbt value.
}
\description{
PendingTicks data (tag 51) holds a list of NBT values for
pending ticks.
}
\details{
\code{get_pending_ticks_data()} and \code{get_pending_ticks_value()} load PendingTicks
data from \code{db}. \code{get_pending_ticks_value()} supports loading
only a single value.

\code{put_pending_ticks_value()} and \code{put_pending_ticks_data()} store
PendingTicks data into \code{db}.
}
