% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{put_data}
\alias{put_data}
\alias{put_value}
\title{Store values to a bedrockdb.}
\usage{
put_data(values, keys, db, writeoptions = NULL)

put_value(value, key, db, writeoptions = NULL)
}
\arguments{
\item{values}{A list of raw values. If \code{keys} is missing, the names of \code{values} will be taken as the keys.}

\item{keys}{A character vector of keys.}

\item{db}{A \code{bedrockdb} object}

\item{writeoptions}{A \code{bedrock_leveldb_writeoptions} object}

\item{value}{A raw vector that contains the information to be written.}

\item{key}{A key that will be used to store data.}
}
\value{
An invisible copy of \code{db}.
}
\description{
Store values to a bedrockdb.
}
\seealso{
\link{delete_data} for removing key-values pairs from \code{db}.

\link{write_data} for applying a batch of \code{put} and \code{delete} operations to \code{db}.
}
