% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbedrock-package.R
\docType{package}
\name{rbedrock-package}
\alias{rbedrock}
\alias{rbedrock-package}
\title{rbedrock: Analysis and Manipulation of Data from Minecraft Bedrock Edition}
\description{
Implements an interface to Minecraft (Bedrock Edition) worlds. Supports the analysis and management of these worlds and game saves.
}
\details{
rbedrock is currently under development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/reedacartwright/rbedrock}
  \item Report bugs at \url{https://github.com/reedacartwright/rbedrock/issues}
}

}
\author{
\strong{Maintainer}: Reed Cartwright \email{racartwright@gmail.com} (\href{https://orcid.org/0000-0002-0837-9380}{ORCID})

Other contributors:
\itemize{
  \item Rich FitzJohn \email{rich.fitzjohn@gmail.com} [contributor]
  \item Christian Stigen Larsen [contributor]
  \item The LevelDB Authors [copyright holder]
}

}
\keyword{internal}
