% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbedrock.R
\docType{package}
\name{rbedrock}
\alias{rbedrock}
\title{rbedrock: A package for the analysis and manipulation of Minecraft: Bedrock Edition worlds.}
\description{
rbedrock: A package for the analysis and manipulation of Minecraft: Bedrock Edition worlds.
}
