% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbt.R
\name{read_nbt}
\alias{read_nbt}
\alias{write_nbt}
\title{Read and Write NBT Data}
\usage{
read_nbt(rawval, max_elements = NULL, simplify = TRUE)

write_nbt(object)
}
\arguments{
\item{rawval}{A \code{raw} vector of binary data to parse.}

\item{max_elements}{Maximum number of elements to parse.}

\item{simplify}{If TRUE, simplifies a list containing a single unnamed \code{nbtnode}.}

\item{object}{A single object of class \code{nbtnode} or a named list of such objects.}
}
\description{
\code{read_nbt} reads NBT data from a \code{raw} vector.

\code{write_nbt} writes a single \code{nbtnode} or a list of \code{nbtnodes} into \code{raw} vector.
}
