% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbt.R
\name{from_rnbt}
\alias{from_rnbt}
\alias{to_rnbt}
\alias{from_rnbt_payload}
\alias{to_rnbt_payload}
\alias{read_rnbt}
\alias{write_rnbt}
\title{Raw Named Binary Tag Format}
\usage{
from_rnbt(x)

to_rnbt(x)

from_rnbt_payload(x, tag)

to_rnbt_payload(x)

read_rnbt(rawdata)

write_rnbt(object)
}
\arguments{
\item{x}{an object}

\item{rawdata}{A \code{raw} vector}

\item{object}{An nbt object or a list of nbt objects}
}
\description{
\code{from_rnbt()} converts \code{rnbt} data to \code{nbt} data.
\code{to_rnbt()} converts \code{nbt} data to \code{rnbt} data.
\code{read_rnbt()} converts a \code{raw} vector to \code{rnbt} data.
\code{write_rnbt()} converts \code{rnbt} data to a \code{raw} vector
}
\details{
\code{rnbt} is a recursive, intermediate data structure that closely
resembles how NBT data is encoded.
}
\keyword{internal}
